+++ shares/models.py
@@ -5,7 +5,7 @@ from circle.models import Circle, CircleMember
 
 class Shares(models.Model):
     circle_member = models.ForeignKey(CircleMember,related_name='shares',  null=False)
-    num_of_shares = models.FloatField(blank=False, null=False, default=0.00)
+    num_of_shares = models.IntegerField(blank=False, null=False, default=0)
 
     class Meta:
         db_table = 'Shares'
@@ -16,7 +16,7 @@ class IntraCircleShareTransaction(models.Model):
     transaction_type = models.CharField(choices=TRANSACTION_TYPE,max_length=8)
     recipient = models.ForeignKey(CircleMember, null=True,related_name='recipient')
     sender = models.ForeignKey(CircleMember, null=True,related_name='sender')
-    num_of_shares = models.FloatField(blank=False, null=False, default=0.00)
+    num_of_shares = models.IntegerField(blank=False, null=False, default=0)
     transaction_desc = models.TextField(max_length=10000, blank=False)
     transaction_time = models.DateTimeField(null=False, auto_now=True)
 
@@ -26,7 +26,7 @@ class IntraCircleShareTransaction(models.Model):
 
 class LockedShares(models.Model):
     shares = models.ForeignKey(Shares,null=False,default=75)
-    num_of_shares = models.FloatField(blank=False, null=False, default=0.00)
+    num_of_shares = models.IntegerField(blank=False, null=False, default=0)
     transaction_description = models.TextField(max_length=1000, blank=False)
     time_of_transaction = models.DateTimeField(null=False, auto_now=True)
     extra_info = models.TextField(max_length=10000, blank=False)
@@ -37,7 +37,7 @@ class LockedShares(models.Model):
 
 class UnlockedShares(models.Model):
     circle_member = models.ForeignKey(CircleMember, null=False)
-    num_of_shares = models.FloatField(blank=False, null=False, default=0.00)
+    num_of_shares = models.IntegerField(blank=False, null=False, default=0)
     transaction_description = models.TextField(max_length=1000, blank=False)
     time_of_transaction = models.DateTimeField(null=False, auto_now=True)
     extra_info = models.TextField(max_length=10000, blank=False)
